+++ as/cNFT/assembly/market.ts
@@ -7,7 +7,6 @@ import { persistent_tokens_royalty } from './models/persistent_tokens_royalty'
 import { persistent_tokens } from './models/persistent_tokens'
 import { XCC_GAS } from '../../utils'
 import { assert_eq_attached_deposit, assert_one_yocto, assert_token_exists, assert_eq_token_owner, assert_not_paused } from './utils/asserts'
-import {internal_nft_is_approved, nft_is_approved} from "./core";
 
 class NftTransferArgs {
     token_id: string
